userId

+++ src/services/contacts.js
@@ -1,7 +1,5 @@
 import contactCollection from '../db/models/Contacts.js';
-
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
-
 import { SORT_ORDER } from '../constants/index.js';
 
 export const getContacts = async ({
@@ -9,13 +7,13 @@ export const getContacts = async ({
   perPage = 10,
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
-
-  const contactsQuery = contactCollection.find();
+  const contactsQuery = contactCollection.find({ userId });
   const contactsCount = await contactCollection
-    .find()
+    .find({ userId })
     .merge(contactsQuery)
     .countDocuments();
 
@@ -36,19 +34,23 @@ export const getContacts = async ({
 //
 
 export const getContactId = (id) => {
-  return contactCollection.findOne({ _id: id });
+  return contactCollection.findOne({ _id: id, userId });
 };
 
 export const addContact = (payload) => contactCollection.create(payload);
 
-export const updateContact = async (_id, payload, option = {}) => {
+export const updateContact = async (_id, payload, userId, option = {}) => {
   const { upsert = false } = option;
 
-  const rawResult = await contactCollection.findOneAndUpdate({ _id }, payload, {
-    new: true,
-    upsert,
-    includeResultMetadata: true,
-  });
+  const rawResult = await contactCollection.findOneAndUpdate(
+    { _id, userId },
+    payload,
+    {
+      new: true,
+      upsert,
+      includeResultMetadata: true,
+    },
+  );
 
   if (!rawResult.value) {
     return null;
@@ -60,5 +62,5 @@ export const updateContact = async (_id, payload, option = {}) => {
   };
 };
 
-export const deleteContactById = (_id) =>
-  contactCollection.findOneAndDelete({ _id });
+export const deleteContactById = (_id, userId) =>
+  contactCollection.findOneAndDelete({ _id, userId });
